fixed slider event not firing bug
When you click a slider, it was setting the canvas up to skip the next
mouse down event. However it was already setup to skip it based on the
type of the input. As a result it would not draw directly after using a
slider.

Event skipping is now removed, fixing the bug.