InfoBar no longer flickers
Before the infobar was hidden everytime a top button was clicked,
even if that top button would end up showing the infobar.

If the infobar was already open, and you clicked on a control that also
used it,
then this would cause it to flicker on the ipad (because it hides and
then shows).

Now it hides the infobar if it's clicked by a button that doesn't use
it.