changed infobar to hide/show using css classes
Before the info bar was manually hidden using inlined CSS,
namely display: none and display: block.

Now this is pushed into a CSS class, so just adding 'sb_show'
displays it, and removing it hides it.

Also changed a load of '_this' variables to 'self',
and some 'null's to 'nil's.