fixed cursor location moving on size change
Before the lastY and lastX locations in the cursor was failing to be
changed,
so this has been fixed.

The BrushCursor also rendered slightly incorrectly before,
where anti-aliasing around the edges would get chopped off.
A few extra pixels have been added to fix this,
and rendering location of square and circle has been tweaked
so they slightly better match up to the actual size.