fixed canvas zooming into the center
Before the canvas would zoom into the top left corner,
due to a previous change.

The zoom code has been simplified, and now zooms in relation
to the center of the canvas. It also has an extra zoom option, allowing
you to just say "zoom into the center" when you update the zoom
value.

There are also some minor changes to the code, replacing calls to
'css' with 'width', 'height', 'show', 'hide' and 'translate'
alternatives.