moved images so they are now under their own folder
Previously, SkyBrush was designed to be under it's own
/skybrush folder. In practice, people want to often have
an 'images', 'css', and 'js' folder in public, and would
want to put SkyBrush there.

The issue is that having images in /images,
heavily pollutes the /images folder for the used
website (as it may have conflicting images).

By adding a folder at /skybrush/images/skybrush, it
allows the file structure to support both, without changes.

Put it under it's own /skybrush folder, or remove the /skybrush,
and split it into folders at /images, /css and /js.