removed 'proto'
The 'proto' fix didn't work for loading SkyBrush
before 'jquery.more', as it was in 'jquery.more' (duh!).

So it has been removed, and instead the command
prototypes are defined when 'newCommands' is
called.

This also allows them to only be setup in a more
lazy way, and avoid this expense, if SkyBrush is not
used.