fixed colour picker not finding painter, when in use
It used to be, that the commands could see
the SkyBrush painter object, when they are created.
This was removed, to help decouple code.

So when the colour picker was in use, it would fail
to access painter.

Painter however was being passed into the pickers
'onDownOnMove' function, and so I just had to add
that parameter in on the colour pickers side.