Fixed a bug with the canvas not updating it's position correctly
The canvas location was being relaid out in Chrome directly,
with no delay. This would break with things which had an animated
change in size, such as when opening the GUI bar.

The canvas now resizes in a lazy way, where you call and it updates in
the future. It also ignores repeated calls, only updating for the most
recent call.