fixed a bug with the shape commands
The constructors for creating the shape commands for prototypes was
wrong.

They must have been using an old interface which caused the code to end
up setting properties to a string value. Before it must have silently
done nothing but now Chrome throws an error. The code has been
corrected.